feat(KLS-107): Project reorganization
