+++ src/components/MediaPlayer.vue
