+++ src/views/exampleOne.vue
@@ -9,7 +9,7 @@
     </router-link>
     <div class="container mx-auto max-w-6xl px-3 my-9">
       <div class="flex justify-center">
-        <ControlAnimations
+        <MediaPlayer
           classes="w-full "
           source="models/strong_chicken/scene.gltf"
           domElement="threejs2"
@@ -23,7 +23,7 @@
 
             backgroundColor: '#9a9a9ab8',
           }"
-        ></ControlAnimations>
+        ></MediaPlayer>
       </div>
       <p>
         Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc vel diam
@@ -82,12 +82,11 @@
 </template>
 
 <script>
-import ControlAnimations from "@/components/controlAnimations.vue";
-
+import MediaPlayer from "@/components/MediaPlayer";
 export default {
   name: "exampleOne",
   components: {
-    ControlAnimations,
+    MediaPlayer,
   },
   props: {
     msg: String,
