--- src/views/exampleTwo.vue
@@ -1,121 +1,0 @@
-<template>
-  <div class="scene relative">
-    <router-link to="/">
-      <button
-        class="bg-blue-400 text-white hover:bg-white hover:text-blue-400 hover:border hover:border-blue-400 px-4 py-2 rounded-xl absolute left-0 m-6"
-      >
-        Home
-      </button>
-    </router-link>
-
-    <canvas id="myCanvas"></canvas>
-  </div>
-</template>
-
-<script>
-import * as THREE from "three";
-import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader";
-import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
-import { DRACOLoader } from "three/examples/jsm/loaders/DRACOLoader";
-//import gltf from "three/examples/js/libs/draco";
-
-export default {
-  name: "exampleTwo",
-  props: {
-    msg: String,
-    scene: undefined,
-  },
-  methods: {
-    init() {
-      const scene = new THREE.Scene();
-
-      const camera = new THREE.PerspectiveCamera(
-        75,
-        window.innerWidth / window.innerHeight,
-        0.1,
-        1000
-      );
-      const canvas = document.getElementById("myCanvas");
-      const renderer = new THREE.WebGLRenderer({
-        canvas: canvas,
-        antialias: true,
-        alpha: true,
-      });
-      const controls = new OrbitControls(camera, renderer.domElement);
-      //controls.enableDamping = true;
-      controls.screenSpacePanning = true;
-
-      //const renderer = new THREE.WebGLRenderer();
-      renderer.setSize(window.innerWidth, window.innerHeight);
-      // renderer.domElement.id = "scene3D";
-
-      /* console.log("scene:", scene);
-      const geometry = new THREE.BoxGeometry(1, 1, 1);
-      const material = new THREE.MeshBasicMaterial({ color: 0x00ff00 });
-      const cube = new THREE.Mesh(geometry, material);
-      scene.add(cube); */
-
-      const loader = new GLTFLoader();
-      const dracoLoader = new DRACOLoader();
-      dracoLoader.setDecoderPath(
-        "https://raw.githubusercontent.com/mrdoob/three.js/dev/examples/js/libs/draco/"
-      );
-      loader.setDRACOLoader(dracoLoader);
-      loader.load(
-        "models/bathroom_interior/scene.gltf",
-        function (gltf) {
-          scene.add(gltf.scene);
-
-          gltf.animations; // Array<THREE.AnimationClip>
-          gltf.scene; // THREE.Group
-          gltf.scenes; // Array<THREE.Group>
-          gltf.cameras; // Array<THREE.Camera>
-          gltf.asset; // Object
-        },
-        // called while loading is progressing
-        function (xhr) {
-          console.log((xhr.loaded / xhr.total) * 100 + "% loaded");
-        },
-        // called when loading has errors
-        function (error) {
-          console.log("An error happened: ", error);
-        }
-      );
-
-      /*  camera.position.z = 5;
-      camera.position.y = 0; */
-      camera.position.set(0, 2, 2);
-      controls.target.set(0, 2, 0);
-
-      //camera.rotation.set(100, 100, 100);
-      controls.update();
-
-      window.addEventListener("resize", onWindowResize, false);
-      function onWindowResize() {
-        camera.aspect = window.innerWidth / window.innerHeight;
-        camera.updateProjectionMatrix();
-        renderer.setSize(window.innerWidth, window.innerHeight);
-        render();
-      }
-      function animate() {
-        requestAnimationFrame(animate); /*
-        cube.rotation.x += 0.01;
-        cube.rotation.y += 0.01; */
-        //renderer.render(scene, camera);
-        render();
-      }
-      animate();
-      function render() {
-        renderer.render(scene, camera);
-      }
-
-      console.log("Scene: ", scene);
-    },
-  },
-  computed() {},
-
-  mounted() {
-    this.init();
-  },
-};
-</script>
